Fix race conditions introduced by statistics

+++ cmd/oplogd/main.go
@@ -12,7 +12,7 @@ var (
 	listenAddr           = flag.String("listen", ":8042", "The address to listen on. Same address is used for both SSE(HTTP) and UDP APIs.")
 	mongoURL             = flag.String("mongo-url", "", "MongoDB URL to connect to.")
 	cappedCollectionSize = flag.Int("capped-collection-size", 104857600, "Size of the created MongoDB capped collection size in bytes (default 100MB).")
-	maxQueuedEvents      = flag.Int("max-queued-events", 100000, "Number of events to queue before starting throwing UDP messages.")
+	maxQueuedEvents      = flag.Uint64("max-queued-events", 100000, "Number of events to queue before starting throwing UDP messages.")
 	password             = flag.String("password", "", "Password protecting the global SSE stream.")
 )
